authorization feature

+++ src/api/notes/handler.js
@@ -16,8 +16,10 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { title = 'untitled', body, tags } = request.payload
-
-      const noteId = await this._service.addNote({ title, body, tags })
+      const { id: credentialId } = request.auth.credentials
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId
+      })
 
       const response = h.response({
         status: 'success',
@@ -38,6 +40,7 @@ class NotesHandler {
         return response
       }
 
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.'
@@ -48,8 +51,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler () {
-    const notes = await this._service.getNotes()
+  async getNotesHandler (request) {
+    const { id: credentialId } = request.auth.credentials
+    const notes = await this._service.getNotes(credentialId)
     return {
       status: 'success',
       data: {
@@ -61,7 +65,11 @@ class NotesHandler {
   async getNoteByIdHandler (request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialId)
       const note = await this._service.getNoteById(id)
+
       return {
         status: 'success',
         data: {
@@ -78,6 +86,7 @@ class NotesHandler {
         return response
       }
 
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.'
@@ -91,10 +100,10 @@ class NotesHandler {
   async putNoteByIdHandler (request, h) {
     try {
       this._validator.validateNotePayload(request.payload)
-      const { title, body, tags } = request.payload
       const { id } = request.params
-
-      await this._service.editNoteById(id, { title, body, tags })
+      const { id: credentialId } = request.auth.credentials
+      await this._service.verifyNoteOwner(id, credentialId)
+      await this._service.editNoteById(id, request.payload)
 
       return {
         status: 'success',
@@ -110,6 +119,7 @@ class NotesHandler {
         return response
       }
 
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.'
@@ -123,6 +133,8 @@ class NotesHandler {
   async deleteNoteByIdHandler (request, h) {
     try {
       const { id } = request.params
+      const { id: credentialId } = request.auth.credentials
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.deleteNoteById(id)
 
       return {
@@ -139,6 +151,7 @@ class NotesHandler {
         return response
       }
 
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.'
